# openapi.yaml
openapi: 3.0.3

info:
  title: RFID Attendance System API
  description: |
    REST API for the RFID employee attendance system.

    **Auth:** Session cookie (`connect.sid`) returned on login.
    Routes marked 🔒 require an active admin session.
    `/api/check` is public (called by the scan page).
  version: 1.0.0
  contact:
    name: RFID Attendance

servers:
  - url: http://localhost:3000
    description: Local dev (change to production URL when deployed)

tags:
  - name: auth
    description: Session management
  - name: employees
    description: Employee registry (🔒 admin only)
  - name: attendance
    description: Card scan and attendance records
  - name: reports
    description: Reporting (🔒 admin only)
  - name: devices
    description: Server-side RFID reader management (🔒 admin only for write)

paths:

  # ── Auth ──────────────────────────────────────────────────────────────────

  /api/login:
    post:
      tags: [auth]
      operationId: login
      summary: Admin login
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [username, password]
              properties:
                username:
                  type: string
                  example: admin
                password:
                  type: string
                  example: admin123
      responses:
        "200":
          description: Login successful — session cookie set
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
                    example: true
                  username:
                    type: string
                    example: admin
        "401":
          $ref: "#/components/responses/InvalidCredentials"

  /api/logout:
    post:
      tags: [auth]
      operationId: logout
      summary: Destroy current session
      security: []
      responses:
        "200":
          description: Session destroyed
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
                    example: true

  /api/me:
    get:
      tags: [auth]
      operationId: getMe
      summary: Check current session state
      security: []
      responses:
        "200":
          description: Session info (always 200 — check loggedIn field)
          content:
            application/json:
              schema:
                type: object
                properties:
                  loggedIn:
                    type: boolean
                  admin:
                    oneOf:
                      - $ref: "#/components/schemas/AdminInfo"
                      - type: object
                        nullable: true
                        description: null when not logged in

  # ── Employees ─────────────────────────────────────────────────────────────

  /api/employees:
    get:
      tags: [employees]
      operationId: listEmployees
      summary: List all employees 🔒
      security:
        - sessionCookie: []
      responses:
        "200":
          description: Array of employees ordered by newest first
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Employee"
        "401":
          $ref: "#/components/responses/Unauthorized"

    post:
      tags: [employees]
      operationId: createEmployee
      summary: Register a new employee 🔒
      security:
        - sessionCookie: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [card_id, name]
              properties:
                card_id:
                  type: string
                  description: Unique RFID card identifier
                  example: "A1B2C3D4"
                name:
                  type: string
                  example: "محمد أحمد"
                title:
                  type: string
                  example: "مهندس برمجيات"
      responses:
        "200":
          description: Employee created
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
                    example: true
                  id:
                    type: integer
                    example: 3
        "400":
          description: card_id or name missing
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "409":
          description: Card ID already registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ── Attendance ────────────────────────────────────────────────────────────

  /api/check:
    post:
      tags: [attendance]
      operationId: checkCard
      summary: Record a card scan (check-in / check-out)
      security: []
      description: |
        Public endpoint called by the scan page. Alternates IN→OUT→IN per employee.

        **Cooldown:** If the same card is scanned within 10 seconds of the previous
        record, the existing record is returned unchanged (`deduplicated: true`).
        This prevents double-logging from sticky cards or rapid re-scans.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [card_id]
              properties:
                card_id:
                  type: string
                  example: "A1B2C3D4"
      responses:
        "200":
          description: Attendance recorded (or deduplicated within cooldown window)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CheckResponse"
        "400":
          description: card_id missing or empty
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Card not registered or employee inactive
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "بطاقة غير معرفة"
                  card_id:
                    type: string
                    example: "UNKNOWN123"

  /api/today:
    get:
      tags: [attendance]
      operationId: getToday
      summary: All attendance records for today 🔒
      description: Returns up to 100 records, newest first. "Today" resolved in server local time.
      security:
        - sessionCookie: []
      responses:
        "200":
          description: Today's attendance log
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AttendanceRecord"
        "401":
          $ref: "#/components/responses/Unauthorized"

  # ── Reports ───────────────────────────────────────────────────────────────

  /api/reports:
    get:
      tags: [reports]
      operationId: getReport
      summary: Attendance report by date range and optional employee filter 🔒
      description: |
        Returns one entry per employee per day: earliest IN and latest OUT.
        Filter by date range (`from`/`to`) and optionally by a single employee.
        Falls back to current month if `from`/`to` are omitted (uses `year`/`month`).
      security:
        - sessionCookie: []
      parameters:
        - name: from
          in: query
          required: false
          schema:
            type: string
            format: date
            example: "2026-06-01"
          description: Start date (YYYY-MM-DD). Use with `to`.
        - name: to
          in: query
          required: false
          schema:
            type: string
            format: date
            example: "2026-06-30"
          description: End date inclusive (YYYY-MM-DD). Use with `from`.
        - name: employee_id
          in: query
          required: false
          schema:
            type: integer
            example: 1
          description: Restrict to one employee. Omit for all employees.
        - name: year
          in: query
          required: false
          schema:
            type: string
            pattern: '^\d{4}$'
            example: "2026"
          description: Fallback if from/to not provided.
        - name: month
          in: query
          required: false
          schema:
            type: string
            pattern: '^\d{2}$'
            example: "06"
          description: Fallback if from/to not provided.
      responses:
        "200":
          description: Report grouped by employee
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Report"
        "400":
          description: Invalid date format or employee_id
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"

  # ── RFID Devices ──────────────────────────────────────────────────────────

  /api/devices:
    get:
      tags: [devices]
      operationId: listDevices
      summary: List connected USB HID devices 🔒
      security:
        - sessionCookie: []
      responses:
        "200":
          description: HID device list and node-hid availability
          content:
            application/json:
              schema:
                type: object
                properties:
                  available:
                    type: boolean
                    description: true if node-hid is installed and usable
                  devices:
                    type: array
                    items:
                      $ref: "#/components/schemas/HidDevice"
        "401":
          $ref: "#/components/responses/Unauthorized"

  /api/devices/status:
    get:
      tags: [devices]
      operationId: getDeviceStatus
      summary: Current server-side reader status (public)
      security: []
      responses:
        "200":
          description: Reader status
          content:
            application/json:
              schema:
                type: object
                properties:
                  available:
                    type: boolean
                  active:
                    type: boolean
                    description: true if a reader is open and listening
                  vendorId:
                    type: integer
                    nullable: true
                  productId:
                    type: integer
                    nullable: true

  /api/devices/select:
    post:
      tags: [devices]
      operationId: selectDevice
      summary: Select a USB HID device as the RFID reader 🔒
      security:
        - sessionCookie: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [vendorId, productId]
              properties:
                vendorId:
                  type: integer
                  example: 1155
                productId:
                  type: integer
                  example: 22352
      responses:
        "200":
          description: Reader opened and settings persisted
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
        "400":
          description: Missing or invalid vendorId/productId
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          description: Could not open HID device
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    delete:
      tags: [devices]
      operationId: deselectDevice
      summary: Stop the server-side RFID reader 🔒
      security:
        - sessionCookie: []
      responses:
        "200":
          description: Reader closed and settings cleared
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
        "401":
          $ref: "#/components/responses/Unauthorized"

  /api/events:
    get:
      tags: [devices]
      operationId: getEvents
      summary: Server-Sent Events stream for real-time scan results (public)
      security: []
      description: >
        Streams newline-delimited JSON events to connected browsers.
        Event types: "status" (reader connected/disconnected) and
        "scan" (card scanned with employee info and IN/OUT type).
      responses:
        "200":
          description: SSE stream (Content-Type text/event-stream)

  /api/reports/monthly:
    get:
      tags: [reports]
      operationId: getMonthlyReportLegacy
      summary: "(Legacy) Redirects to /api/reports 🔒"
      description: Kept for backwards compatibility. Redirects 307 to `/api/reports` with same query params.
      security:
        - sessionCookie: []
      parameters:
        - name: year
          in: query
          schema: { type: string, example: "2026" }
        - name: month
          in: query
          schema: { type: string, example: "06" }
      responses:
        "307":
          description: Temporary redirect to /api/reports
        "401":
          $ref: "#/components/responses/Unauthorized"

# ── Components ────────────────────────────────────────────────────────────────

components:

  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: connect.sid
      description: Set automatically by POST /api/login

  responses:
    Unauthorized:
      description: No active admin session
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
                example: "UNAUTHORIZED"
    InvalidCredentials:
      description: Wrong username or password
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
                example: "بيانات الدخول غير صحيحة"

  schemas:

    ErrorResponse:
      type: object
      properties:
        error:
          type: string

    HidDevice:
      type: object
      properties:
        vendorId:
          type: integer
          example: 1155
        productId:
          type: integer
          example: 22352
        path:
          type: string
          example: "/dev/hidraw0"
        manufacturer:
          type: string
          example: "ACS"
        product:
          type: string
          example: "ACR122U PICC Interface"
        usagePage:
          type: integer
          example: 1
        usage:
          type: integer
          example: 6

    AdminInfo:
      type: object
      properties:
        id:
          type: integer
        username:
          type: string

    Employee:
      type: object
      properties:
        id:
          type: integer
          example: 1
        card_id:
          type: string
          example: "A1B2C3D4"
        name:
          type: string
          example: "محمد أحمد"
        title:
          type: string
          example: "مهندس برمجيات"
        active:
          type: integer
          enum: [0, 1]
          example: 1
        created_at:
          type: string
          example: "2026-06-13 10:00:00"

    CheckResponse:
      type: object
      properties:
        ok:
          type: boolean
          example: true
        type:
          type: string
          enum: [IN, OUT]
          example: IN
        employee:
          type: object
          properties:
            id:
              type: integer
            name:
              type: string
            title:
              type: string
        time:
          type: string
          example: "6/13/2026, 10:00:00 AM"
        deduplicated:
          type: boolean
          description: Present and true when within the 10-second cooldown window

    AttendanceRecord:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        title:
          type: string
        card_id:
          type: string
        type:
          type: string
          enum: [IN, OUT]
        created_at:
          type: string
          example: "2026-06-13 10:00:00"

    Report:
      type: object
      properties:
        from:
          type: string
          format: date
          example: "2026-06-01"
        to:
          type: string
          format: date
          example: "2026-06-30"
        employees:
          type: array
          items:
            type: object
            properties:
              emp_id:
                type: integer
              name:
                type: string
              title:
                type: string
              card_id:
                type: string
              total_days:
                type: integer
                description: Number of distinct days with at least one scan
              days:
                type: array
                items:
                  type: object
                  properties:
                    day:
                      type: string
                      example: "2026-06-13"
                    first_in:
                      type: string
                      nullable: true
                      example: "08:02:11"
                    last_out:
                      type: string
                      nullable: true
                      example: "17:01:44"